# binary/stream/reader: Fast-path offsetReader skips