--- a/protocol/binary/stream_reader.go
+++ b/protocol/binary/stream_reader.go
@@ -66,10 +66,29 @@ type StreamReader struct {
 
 	reader io.Reader
 	buffer [8]byte
+
+	// discard is points to either discardOffset or discardStream based on
+	// the implementation of the io.Reader we're using.
+	discard func(int64) error
+
+	// These are bound versions of the discardStream and discardOffset
+	// methods on the StreamReader. Putting them here ensures that we don't
+	// cause an alloc when we do "sr.discard = sr.discardOffset".
+	_discardStream func(int64) error
+	_discardOffset func(int64) error
+
+	// This field is set only if the wrapped reader is an offset reader.
+	// DO NOT USE if you're not discardOffset.
+	_or *offsetReader
 }
 
 var streamReaderPool = sync.Pool{
-	New: func() interface{} { return &StreamReader{} },
+	New: func() interface{} {
+		sr := new(StreamReader)
+		sr._discardOffset = sr.discardOffset
+		sr._discardStream = sr.discardStream
+		return sr
+	},
 }
 
 // NewStreamReader fetches a StreamReader from the system that will write
@@ -77,13 +96,21 @@ var streamReaderPool = sync.Pool{
 //
 // This StreamReader must be closed using `Close()`
 func NewStreamReader(r io.Reader) *StreamReader {
-	streamReader := streamReaderPool.Get().(*StreamReader)
-	streamReader.reader = r
-	return streamReader
+	sr := streamReaderPool.Get().(*StreamReader)
+	sr.reader = r
+	sr.discard = sr._discardStream
+	if or, ok := r.(*offsetReader); ok {
+		// If we're wrapping an offsetReader, we can skip bytes much
+		// more efficiently.
+		sr._or = or
+		sr.discard = sr._discardOffset
+	}
+	return sr
 }
 
 func returnStreamReader(sr *StreamReader) {
 	sr.reader = nil
+	sr._or = nil
 	streamReaderPool.Put(sr)
 }
 
@@ -98,7 +125,12 @@ func (sr *StreamReader) read(bs []byte) (int, error) {
 	return n, err
 }
 
-func (sr *StreamReader) discard(n int64) error {
+func (sr *StreamReader) discardOffset(n int64) error {
+	sr._or.offset += n
+	return nil
+}
+
+func (sr *StreamReader) discardStream(n int64) error {
 	_, err := io.CopyN(ioutil.Discard, sr.reader, n)
 	if err == io.EOF {
 		// All EOFs are unexpected when streaming